refactored ImageSeries into more classes to avoid mutating state
Previously, the data attribute of ImageSeries could take on many different types, including NXField, NXVirtualField, and np.ndarray. Certain operations would mutate the type, such that member functions no longer worked.

In order to avoid this, I created an abstract base class defining a minimum set of methods, and derived several classes from it, each having a different relationship with the underlying data: LazyImageSeries, VirtualImageSeries, and InMemoryImageSeries. Each derived class has methods that work only for its particular data type, so it is no longer necessary to keep track of the state of the object during computations.

In order to not change the previous ImageSeries interface, a wrapper ImageSeries class was created which gives a deprecation warning when instantiated. The tests are unchanged -- they still use the ImageSeries interface -- and all tests are passing.

In addition, for a different application, I created a SparseImageSeries object prototype that is included in this commit.